Move indexer init to fulltext index package
